Student Dashboard Metadata and Fixes
